Put the switch & lamp delays on defines
The last version used the clock speed to calculate the delays for lamps & switches. It used the floating point library, and for some people created a horrible bug where solenoids would fire randomly and the ISR malfunctioned. Putting them back on defines puts the responsibility on the programmer for now to update their delays based on the clock speed. In the future, this delay could be calculated by sampling the clock for a second to figure out the frequency.