Removed 0 as a possible value for WizardSwitchReward
I was just reading a comment that a value of zero for WizardSwitchReward "makes absolutely no sense." You know, in general, I'm a big fan of code that's completely configurable even if some of those configuration cases make no sense to casual inspection. I think what the poster was suggesting was that a *default* value of zero makes no sense. That, I might agree with, except there are no real default values with these parameters. When the Arduino is plugged in for the first time, it's in a completely unknown state. I can't tell the difference between a valid an invalid value in the EEPROM except if it's like 0xFFFFFFFF or whatever. I do check for that. But zero was a configurable value so it wasn't invalid.  
Anyway, I took it out. The game is now less configurable. Value added by this commit = 0.